Simplify how the ReceiveTab textblock gets focused.
