+++ WalletWasabi.Gui/Behaviors/FocusBehavior.cs
@@ -1,0 +1,46 @@
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Xaml.Interactivity;
+using System;
+using System.Reactive.Disposables;
+
+namespace WalletWasabi.Gui.Behaviors
+{
+    class FocusBehavior : Behavior<Control>
+    {
+        private CompositeDisposable _disposables = new CompositeDisposable();
+        private Control _attachedControl;
+
+        static readonly AvaloniaProperty<bool> IsFocusedProperty = 
+			AvaloniaProperty.Register<FocusBehavior, bool>(nameof(IsFocused), defaultBindingMode: Avalonia.Data.BindingMode.TwoWay);
+
+        public bool IsFocused
+		{
+			get => GetValue(IsFocusedProperty);
+			set => SetValue(IsFocusedProperty, value);
+		}
+
+        protected override void OnAttached()
+        {
+            base.OnAttached();
+
+            AssociatedObject.AttachedToLogicalTree += (sender, e) =>
+            {
+				_disposables.Add(this.GetObservable(IsFocusedProperty).Subscribe(focused =>
+				{
+					if (focused)
+					{
+						AssociatedObject.Focus();
+					}
+				}));
+            };
+        }
+
+        protected override void OnDetaching()
+        {
+            base.OnDetaching();
+
+            _disposables.Dispose();
+        }
+    }
+}
